chore: split memory_selection tests into dedicated file
